v1.12

+++ src/main/java/com/ahd/api/company/CompanyService.java
@@ -1,16 +1,20 @@
 package com.ahd.api.company;
 
 
-import com.ahd.api.almaktab.AlmaktabModel;
+
+import com.ahd.api.exception.BadRequest;
+import com.ahd.api.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -22,7 +26,6 @@ public class CompanyService {
     public CompanyModel getCompany(int id) {
         return companyRepository.findById(id).get();
     }
-
 
     public List<CompanyModel> getCompanyList(Integer pageNo, Integer pageSize, String sortBy) {
         Pageable paging = PageRequest.of(pageNo, pageSize, Sort.by(sortBy));
@@ -42,8 +45,19 @@ public class CompanyService {
         return companyRepository.save(company);
     }
 
-    public void deleteCompany(int id) {
-        companyRepository.deleteById(id);
+    public CompanyModel deleteCompany(int id) {
+        try {
+            if (companyRepository.existsById(id)) {
+                CompanyModel companyModel = companyRepository.findById(id).get();
+                companyRepository.deleteById(id);
+                return companyModel;
+            }
+        }catch (NoSuchElementException exception) {
+            throw new NotFoundException(String.format("Not Found this id [%s]", id));
+        }catch (HttpClientErrorException.BadRequest exception){
+            throw new BadRequest("illegal request");
+        }
+        throw new NotFoundException(String.format("Not Found this id [%s]", id));
     }
 
     public CompanyModel findByName(String name) {
