+++ src/main/java/com/ahd/api/delegates/DelegatesService.java
@@ -1,15 +1,20 @@
 package com.ahd.api.delegates;
 
 
+import com.ahd.api.company.CompanyModel;
+import com.ahd.api.exception.BadRequest;
+import com.ahd.api.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -39,8 +44,19 @@ public class DelegatesService {
         return delegatesRepository.save(delegates);
     }
 
-    public void deleteDelegates(int id) {
-        delegatesRepository.deleteById(id);
+    public DelegatesModel deleteDelegates(int id) {
+        try {
+            if (delegatesRepository.existsById(id)) {
+                DelegatesModel companyModel = delegatesRepository.findById(id).get();
+                delegatesRepository.deleteById(id);
+                return companyModel;
+            }
+        }catch (NoSuchElementException exception) {
+            throw new NotFoundException(String.format("Not Found this id [%s]", id));
+        }catch (HttpClientErrorException.BadRequest exception){
+            throw new BadRequest("illegal request");
+        }
+        throw new NotFoundException(String.format("Not Found this id [%s]", id));
     }
 
     public DelegatesModel findByName(String name) {
